refactor: controller 와 service 계층 모두에서 사용하는 dto 이동
