+++ src/main/java/chess/service/dto/PasswordRequestDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.service.dto;
 
 public class PasswordRequestDto {
