+++ src/test/java/chess/dao/JdbcGameDaoTest.java
@@ -3,7 +3,7 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.controller.dto.ChessRequestDto;
+import chess.service.dto.ChessRequestDto;
 import chess.service.dto.GameDto;
 import chess.service.dto.GameStatusDto;
 import org.junit.jupiter.api.AfterEach;
